Search Bar + Toggle Campus +  Test
Started the implementation of the Search Bar to show as on top of the maps page. Currently, the button is shown next is to be able to search for location from it and show the results from Google Places API. Toggle button has also been moved to the front maps page so it's easier to access to switch around campus. Test has been implemented to ensure it the toggle button indeed exists.