Upgraded Gradle
Simply upgraded the gradle file.

For those of you might get an error due to this. Please do the following on Android studio:

File -> Project Structure -> Project

Choose the latest Gradle version and hit apply. If you get an option to sync your gradle file then please click on it.