Updated Test file to work
The map fragment test basically ensures that it is in the correct view, has the search bar on the view, has the toggle button, when pressed the toggle button it is indeed checked and has the text SGW, when pressed again it is indeed unchecked and has the text LOY. Additionally, the toggle button is changed. It was initially a picture that had the text embedded to it. Now it's typed via xml.