Searching + Transportation mode
Additions:
- Set up 2 search bars (a main one), one for destination) on the top of the screen
- Moved transporation mode radio buttons from bottom sheet to top of the screen
- Set up the search programmically to work with the Google Places Autocomplete on both search bars
- Clicking on directions button starts navigation from current location, to the selected building
with the selected transporation mode (it also adds the names in the search field)
- The MaterialSearchBar library was used for the search bars

- TODO: navigation starts off right away after clicking directions (there should be some sort of confirmation
+ to allow the user to change their text fields as well)

- TODO: improve UI (ie: top layout margins are hardcoded)

- Regression : clicking on a search result doesnt go to location anymore
(it is simple though, must set up prediction click listener and search result listener to handle it)

- TODO: resolve search bar bugs
The MaterialSearchBar library has some inherit bugs. For example, sometimes the suggestings
list (where the Autocomplete results are shown) doesnt get hidden even though it has been
called to do so explicitly. However as it is an open source library it is possible to
make changes to the library to fit our needs and import it manually after.

-TODO: limit the search results to Greater Montreal Area (find the rectangulars bound for this)