Transit mode now show detailed bus/metro instructions
Now under instructions, the transit mode shows detailed instructions as to which bus/metro to take, how many stops it takes and which bus/metro stop to get off at.

Additionally, for walking mode a step by step instructions is displayed along with the distance and time for each steps.

What to do? This part needs heavy formatting. In other words, as a team we have to establish what exactly to show as instructions based on the mode of transportation chosen.