Refactor Map Initialization in MapFragment
The initialization of the google map used to be done in the mapFragment, however, its more appropriate that such logic be done using the mapViewModel. For this commit, a Map class was added which will function as the map model. This model will be used to initialize the map. The mapFragement now receives the map from the mapViewModel which depend on the newly created map model. The mapViewModel is also responsible of drawing the polygons on Concordia's buildings when the map is initialized and before its passed to the mapFragment. Finally, as recommended, gradle build has been updated to version 3.6.1