Issue #5: Smaller initial buildings icons
The initials of the buildings have been reduced in font size (20) and are now in Gill Sans MT font, the same font used for the Campus Compass logo.

This is a major improvement to the UI as the markers do not look huge anymore.

NEXT-UP:
- Find a way to either cluster markers or make the markers disappear, as map zoom level increases (zoom out)
- Unit tests 
- UI tests for highlighting and markers,
possibly using UIAutomator instead of Espresso as per this StackOverflow answer
https://stackoverflow.com/questions/29924564/using-espresso-to-unit-test-google-maps