Issue #5: Observer pattern to make markers disappear
MAJOR:
- Implemented Observer pattern (small variation) to make initials on buildings "disappear"
- Building is a concrete observer, it implements Observer interface
Variation: Subject is an interface as it is implemented by MapFragment, the decision for this is because notifyObservers() must be used in the method onCameraIdle() which is an overridden method from the OnCameraIdleListener class, which in turn is implemented by MapFragment

MINOR:
-Removed app.Application import from MapRepository class as it is not used
- Moved building constants to a companion object

How it works:
Buildings are observers and they are attached (registered) in the onMapReady() method of MapFragment. Only the buildings which have a center location other than [0,0], meaning they have a marker, are registered as observers. 
When the camera is idle, the observers are notified. 
The update method in building class sets the visibility of the marker depending on the map's camera zoom level (which is passed through the method in the superclass -> update method of Subject)

Functionality:
If you zoom out (pinch gesture) far enough, building initials should disappear. 
This is a major UI improvement because if the markers didn't disappear, they would become bigger and clutter the map as you zoom out farther. Prior to this commit, the user could tap on any of those markers and pop up the bottom sheet.