Issue #5: Add places id to buildings
A place is on Google Maps is identified by its place id. This will be useful for the Search future, when a user searches a building on Concordia.

Unfortunately, not all buildings on Google maps are in the database. I submitted requests for a few of them (mainly Loyola). When I get the approval from Google Maps team that these places are updated, I shall update the json

*Places with "place_id" set as "None" will be discarded in the code -> we will have another custom built way of identifying those buildings.