Refactoring and fix
I refactored the functionality for setting the svg out of 2 methods into
a new method.

This ended up fixing a bug where the app would crash when generating a
random path. This is because I had forgot to change the functionality
for setting the svg image in the onclicklistener, and it would fail as
it was still referencing things that didn't exist anymore