Issue #5: Full UI test for building's Initials Markers
MAJOR:
- Changed coordinates for when the camera moves to SGW / Loyola campus. This is to include most buildings.
- UI test for all buildings' markers with the exception of (Samuel Bronfman, SB) as it is unreachable using the APIs. The only swipe left method does not allow to specify an amount of of pixels to swipe, it is a full swipe. Therefore when swiping right on the map, the side drawer is always brought up,

Note: The orders in which the initial markers are clicked is essential, it must be respected! This is due to the fact that the UI object has a huge hit box. If the map is zoomed out too much, the markers that are next to each other overlap with their hitboxes , thus messing up the test.

MINOR:
- Fixed toggle button in MapFragment, as the map initially displays downtown, the logic in the toggle button needed to be inverted (coordinates).
- Fixed assert in MapResositoryTest