Import Optimized + Fixed app state + Mystery of OK resolved
Import Optimized: As usual, replaced .* with their respective specific imports.

Android Allowing Backup: For some reason, some of our devices were keeping states of the application even after clearing it's storage and uninstalling it.
Turns out, start Android 6.0 API level 23, Android now allow Auto backup of states. This needed to be set to false to prevent device from performing auto backup of these states.

Mystery of OK: For some reason different devices were showing OK vs OKAY as the text of the button. Initially the button was called OKAY but was renamed to OK. Could be the text of the button was being cached but regardless, in order for all of us to be consistent and have the same representation of the application in order to run the same set of tests. This specific text has been hard coded as OK.