Classroom Highlighting
I now provide the functionality that if you click on a classroom when searching, it will take you to the floor fragment and display the classroom as highlighted.

In making this change, I refactored the functionality in reading the svg file, as there was a lot of functionality in common with the new highlight method, so I took the opportunity to create a template method from it.