Adding waypoints to ShuttleBus routes and accurate route times
When the user selects the shuttle transporation method, two waypoints will be set.
The wapoints are determined by which is the nearest shuttle bus stop to the origin location.
For example if the closest shuttle bus stop is the sgw one, it will generate a path from
origin -> sgw campus -> loy campus -> destination.
Later I will override the shown steps to be more relevant to a shuttle trip.