Making nav interactions
I set up the interactions in the floor fragment and map fragment, so that you can now click on a button to indicate that you have arrived and it will take you to the next step.

It works in both directions, so from within a building to outside, and from outside to within a building.