Adding a condition to generate walking instructions when the selected transportation method is shuttle
If the selected mode is shuttle, we want to show walking instructions towards and from the bus stops.
The instructions when we are inside the shuttle bus will be overriden to display that fact.
A bad side effect to this change is that if the distance is too far to realistically walk to, it will still show walking instructions.
The displayed time is also unaffected by this. Meaning, it will only the show the time of the actual shuttle trip (ignoring the walking)
To remedy this, it is possible to pass the previously calculated distance between the origin and the bus stop to the generateDirections()
We would then check, if the distance is above a certain threshold, we would generate directions using other transportation methods
Wether the chosen method is biking, driving or transit is unclear how that will be decided or if it is even required as only walking directions were requested.
This is due to the assumption that the shuttle option will mostly be picked when in or around Concordia, where walking is practical.