fixing bug in highlighting
I fixed a bug in highlighting the buildings that I had created when I had refactored out the portion of the code that iterates to create an svg element.

It was essentially missing the ability to append on each line when it had encountered the svg element. I had not considered this and fixed it.