Fix getting stuck in indoor nav
I just fixed the getting stuck in the indoor nav when going from outside to inside.

I also added the ability to go back between fragments when navigating and to be shown the previous step of the navigation. I needed to add this to fix the bug, or else it would have been a botched job.