Fixed issue with transit directions steps stopping midway into navigation
The reason it stopped was because the catch block was triggered due to an exception.
Before the try-catch was inside the for-loop even after an exception it would move on to the next interation regardless. After refactoring, the try-catch was wrapped around the for-loop instead.
I simply put back the try-catch inside the loop and changed the parsing order to not add coordinates if there are no instructions available.