Fixed bugs related to processing new maps
There are some paths that use the cubic bezier curve, and we didn't have a way to handle that before, but now I added some functionality to handle that.

The next bug was that paths were not able to be drawn in certain corners of the map because of the way I define whether a point is within a path element or not. I modified it to be precise