Disabling Shuttle Bus navigation for invalid routes
A route is invalid if the exit shuttle bus stop is more than 3km away from your set destination.
For example, if you set your destination as the Loyola Stinger Dome and you must take the shuttle from Loyola,
you will not be able to begin shuttle navigation as the Stinger Dome is too far from the SGW shuttle bus stop.
When a shuttle bus route is determined to be invalid, its radio button is disabled and the time shown will be reset to '-'.
If the user has the shuttle radio button selected(checked) and later picks an invalid route,
it will uncheck the shuttle button and check the driving one by default.