#55 UI Test for connectivity
MAJOR:
- Updated SplashFragmentTest.kt to include a test for network connectivity.

As of Android 10, wifiManager and was deprecated. It is not possible to turn on/off Wi-Fi on apps using that API. A work-around is to invoke the settings panel with an intent. I wrote the test to accomodate both test cases (android devices with Android 10 and devices with lower versions).
In the test, Wi-Fi will be turned on then off, and vice versa, whether the device had wi-fi turned on or off initially. This is to make sure that the code is fully covered in all test cases.

MINOR:
- Removed a line from MarkerTest.kt as it is not needed

TESTED ON:
- Physical Google Pixel 2 XL - API 29 (Android 10)
- Google Pixel 2 XL Emulator - API 26 (Android 8.0)
- Nexus 5 Emulator - API 25 (Android 7.1)

Test cases:
- Device has Wi-Fi ON, initially
-Device has Wi-Fi OFF, initially

All cases pass on all test devices.