Fixed crash when navigating with indoor locations and going back to the main map through any other feature
If in indoor navigation mode and the user opens Explore or MyPlaces and clicks on a location to view it on the map, the app would crash.
The crash was due to the us switching to floor_fragment without feedback from the 'go indoors button'
The call to navigate to floor_fragment was inside the observe navigation method which triggered it.