Removing TODO's
I don't know why it mentions to bring the camera to the current location when showing a path (I'm guessing that the one who put it there meant the start location), but we decided that if the user decides to leave the fragment and start looking at other things, that we wouldn't force the user to follow the path.

The next todo mentions adding events to the buildings' information,  but we don't even consider that in our road map, so I will remove it.